chore: Implement graph

+++ src/route.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct Route {
     pub to: String,
-    pub distance: u64,
+    pub distance: i32,
 }
